13 - Advanced Generative AI Capstone/pyproject.toml
[project]
name = "advanced-gen-ai-2025"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "chromadb>=1.0.7",
    "django-environ>=0.12.0",
    "pandas>=2.2.3",
    "pydantic-ai>=0.1.6",
    "pypdf2>=3.0.1",
    "sentence-transformers>=4.1.0",
    "streamlit>=1.44.1",
]
